refactor: seed.js now points to bulk data
