+++ script/seed.js
@@ -1,13 +1,47 @@
 'use strict'
-
+const faker = require('faker')
 const db = require('../server/db')
 const {User, Product, Order} = require('../server/db/models')
+
+//configure bulk seeding parameters here
+const NUM_PRODUCTS = 200 //>=3
+const NUM_USERS = 100 //In addition to 3 default users
+const USERS_WITH_ORDERS = 50 //Probability user has an order 0 - 100
+const ORDERS_FULFILLED = 50 //Probability order is fulfilled 0 - 100
+const MAX_LINES_PER_ORDER = 10
+const MAX_QTY_PER_LINE = 8
+const MIN_PRICE = 5
+const MAX_PRICE = 25
+
+const fakeUser = () => {
+  return {
+    email: faker.internet.email(),
+    password: '123',
+    name: faker.name.findName(),
+    address: faker.address.streetAddress(),
+    city: faker.address.city(),
+    state: faker.address.stateAbbr(),
+    zip: faker.address.zipCode(),
+  }
+}
+
+const fakeProduct = (i) => {
+  let image = '/images/candle-' + ((i + 30) % 31) + '.jpg'
+  let prodName = faker.commerce.productAdjective() + ' Candle'
+  return {
+    name: prodName,
+    description: faker.lorem.sentences(2),
+    imageUrl: image,
+    price: Math.floor((MAX_PRICE + 1 - MIN_PRICE) * Math.random() + MIN_PRICE),
+  }
+}
 
 async function seed() {
   await db.sync({force: true})
   console.log('db synced!')
 
-  const users = await Promise.all([
+  //create three default users
+  const defaultUsers = await Promise.all([
     User.create({
       email: 'cody@email.com',
       password: '123',
@@ -15,7 +49,7 @@ async function seed() {
       address: '145 Main Street',
       city: 'Chicago',
       state: 'IL',
-      zip: '60080'
+      zip: '60080',
     }),
     User.create({
       email: 'murphy@email.com',
@@ -24,79 +58,7 @@ async function seed() {
       address: '335 Lakeshore Dr',
       city: 'Chicago',
       state: 'IL',
-      zip: '60080'
-    }),
-    User.create({
-      email: 'elena@email.com',
-      password: '123',
-      name: 'Elena King',
-      address: '157 Lions Ct',
-      city: 'Lake Zurich',
-      state: 'IL',
-      zip: '60047'
-    }),
-    User.create({
-      email: 'sidney@email.com',
-      password: '123',
-      name: 'Sidney Hicks',
-      address: '157 Lions Ct',
-      city: 'Chicago',
-      state: 'IL',
-      zip: '60047'
-    }),
-    User.create({
-      email: 'byron@email.com',
-      password: '123',
-      name: 'Byron Hawkins',
-      address: '2948 Victoria Ln',
-      city: 'Allentown',
-      state: 'PA',
-      zip: '18101'
-    }),
-    User.create({
-      email: 'jenny@email.com',
-      password: '123',
-      name: 'Jenny Howard',
-      address: '8090 Elm St',
-      city: 'Marshall',
-      state: 'VA',
-      zip: '20116'
-    }),
-    User.create({
-      email: 'susie@email.com',
-      password: '123',
-      name: 'Susie Hanson',
-      address: '1252 Oak St',
-      city: 'Berger',
-      state: 'MO',
-      zip: '63014'
-    }),
-    User.create({
-      email: 'aubree@email.com',
-      password: '123',
-      name: 'Aubree Jordan',
-      address: '2708 Victoria Ln',
-      city: 'Allentown',
-      state: 'PA',
-      zip: '18101'
-    }),
-    User.create({
-      email: 'greg@email.com',
-      password: '123',
-      name: 'Greg Parker',
-      address: '8400 Elm St',
-      city: 'Marshall',
-      state: 'VA',
-      zip: '20116'
-    }),
-    User.create({
-      email: 'lisa@email.com',
-      password: '123',
-      name: 'Lisa Carroll',
-      address: '3 Maple St',
-      city: 'Berger',
-      state: 'MO',
-      zip: '63014'
+      zip: '60080',
     }),
     User.create({
       email: 'admin@email.com',
@@ -106,146 +68,72 @@ async function seed() {
       city: 'Chicago',
       state: 'IL',
       zip: '60080',
-      admin: true
-    })
-    // User.create({email: 'murphy@email.com', password: '123'}),
-  ])
-  const products = await Promise.all([
-    Product.create({
-      name: 'Glitter Delight',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-01.jpg',
-      price: 5
+      admin: true,
     }),
-    Product.create({
-      name: 'Lavender Fields',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-02.jpg',
-      price: 15
-    }),
-    Product.create({
-      name: 'Sea Breeze',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-03.jpg',
-      price: 10
-    }),
-    Product.create({
-      name: 'Alpine Spice',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-04.jpg',
-      price: 10
-    }),
-    Product.create({
-      name: 'Spa Sensation',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-05.jpg',
-      price: 25
-    }),
-    Product.create({
-      name: 'Linen Moments',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-06.jpg',
-      price: 17
-    }),
-    Product.create({
-      name: 'Ruby Blush',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-07.jpg',
-      price: 10
-    }),
-    Product.create({
-      name: 'Cinnamon Spice',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-08.jpg',
-      price: 13
-    }),
-    Product.create({
-      name: 'Lily Relaxation',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-09.jpg',
-      price: 16
-    }),
-    Product.create({
-      name: 'Harvest Moon',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-10.jpg',
-      price: 21
-    }),
-    Product.create({
-      name: 'Koala Kandle',
-      description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Et tortor at risus viverra adipiscing at in tellus.',
-      imageUrl: '/images/candle-11.jpg',
-      price: 8
-    })
   ])
 
-  let orders = []
-  for (let i = 0; i < 10; i++) {
-    orders.push(await Order.create({userId: users[i].id, isFulfilled: false}))
+  //Create bulk products
+  let products = []
+  for (let i = 1; i <= NUM_PRODUCTS; i++) {
+    products.push(await Product.create(fakeProduct(i)))
   }
 
-  await orders[0].addProduct(products[0], {
-    through: {quantity: 1, subtotal: 1 * products[0].price}
+  //Create default orders for Cody & Murphy (i.e. the default users)
+  let defaultOrders = []
+  for (let i = 0; i < 2; i++) {
+    defaultOrders.push(
+      await Order.create({userId: defaultUsers[i].id, isFulfilled: false})
+    )
+  }
+
+  await defaultOrders[0].addProduct(products[0], {
+    through: {quantity: 1, subtotal: 1 * products[0].price},
   })
 
-  await orders[1].addProduct(products[0], {
-    through: {quantity: 2, subtotal: 2 * products[0].price}
+  await defaultOrders[1].addProduct(products[0], {
+    through: {quantity: 2, subtotal: 2 * products[0].price},
   })
 
-  await orders[1].addProduct(products[1], {
-    through: {quantity: 1, subtotal: 1 * products[1].price}
+  await defaultOrders[1].addProduct(products[1], {
+    through: {quantity: 1, subtotal: 1 * products[1].price},
   })
 
-  await orders[1].addProduct(products[2], {
-    through: {quantity: 1, subtotal: 1 * products[2].price}
+  await defaultOrders[1].addProduct(products[2], {
+    through: {quantity: 1, subtotal: 1 * products[2].price},
   })
 
-  await orders[1].addProduct(products[3], {
-    through: {quantity: 3, subtotal: 3 * products[3].price}
-  })
+  //Create bulk users (in addition to default users)
+  let users = []
+  for (let i = 1; i <= NUM_USERS; i++) {
+    users.push(await User.create(fakeUser()))
+  }
 
-  await orders[1].addProduct(products[4], {
-    through: {quantity: 2, subtotal: 2 * products[4].price}
-  })
-
-  await orders[2].addProduct(products[5], {
-    through: {quantity: 1, subtotal: 1 * products[5].price}
-  })
-
-  await orders[2].addProduct(products[6], {
-    through: {quantity: 1, subtotal: 1 * products[6].price}
-  })
-
-  await orders[3].addProduct(products[6], {
-    through: {quantity: 1, subtotal: 1 * products[6].price}
-  })
-
-  await orders[4].addProduct(products[6], {
-    through: {
-      quantity: 1,
-      subtotal: 1 * products[6].price
+  //Create bulk orders - fulfilled and unfulfilled, random number line items, random qty
+  for (let i = 0; i < users.length; i++) {
+    //determine if user will have an order
+    if (Math.random() < USERS_WITH_ORDERS / 100) {
+      let orderFulfilled = Math.random() < ORDERS_FULFILLED / 100 //determine if order isFulfilled
+      let order = await Order.create({
+        userId: users[i].id,
+        isFulfilled: orderFulfilled,
+      })
+      let numberLines = Math.ceil(MAX_LINES_PER_ORDER * Math.random()) //determine how many order lines
+      for (let j = 1; j <= numberLines; j++) {
+        let qty = Math.ceil(MAX_QTY_PER_LINE * Math.random()) //determiine qty per line item
+        let ranProdInd = Math.floor(NUM_PRODUCTS * Math.random()) //random product selection
+        let linePrice = qty * products[ranProdInd].price
+        await order.addProduct(products[ranProdInd], {
+          through: {quantity: qty, subtotal: linePrice},
+        })
+      }
     }
-  })
+  }
 
   console.log(`seeded ${users.length} users`)
   console.log(`seeded ${products.length} products`)
   console.log(`seeded successfully`)
 }
 
-// We've separated the `seed` function from the `runSeed` function.
-// This way we can isolate the error handling and exit trapping.
-// The `seed` function is concerned only with modifying the database.
 async function runSeed() {
   console.log('seeding...')
   try {
@@ -260,12 +148,8 @@ async function runSeed() {
   }
 }
 
-// Execute the `seed` function, IF we ran this module directly (`node seed`).
-// `Async` functions always return a promise, so we can use `catch` to handle
-// any errors that might occur inside of `seed`.
 if (module === require.main) {
   runSeed()
 }
 
-// we export the seed function for testing purposes (see `./seed.spec.js`)
 module.exports = seed
